feature: use list of lists to represent columns instead of dict
